Another non-working commit: circular dependency
This book is making me work very, very hard for not very
much knowledge. I'm 90 pages into it, and haven't really
gotten a payoff yet, other than honing my attention to detail.
While increasing attention to detail is always a skill worth
mastering, it would be nice to have at least some payoff
along the way.

One aspect of this style of exposition is that the test
suite doesn't always pass. I'll implement a feature in
one section of the book, push that test to passing, then
implement another section to passing, and so on, until
the feature is fully implemented. The running the entire
test suite indicates failing tests. Commits where the entire
test suite isn't passing make it more difficult to use git
bisect to chase down problems.

Given that some of the features require 8-10 pages of exposition
to get a single test to pass, I wonder if the tests aren't
granular enough. They could certainly be more granular for me,
would help me learn the material better.